sales return complete and some minor changes in the functions

+++ database/migrations/2020_07_29_113732_create_purchase_order_products_table.php
@@ -16,7 +16,7 @@ class CreatePurchaseOrderProductsTable extends Migration
         Schema::create('purchase_order_products', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('purchase_order_id');
-            $table->unsignedInteger('producto_id');
+            $table->unsignedBigInteger('producto_id');
             $table->decimal('required_qty',9,2);
             $table->decimal('price',9,2);
             $table->decimal('accept_qty',9,2)->default('0.00');
